Moved a couple of classes about
